ludus P3: T1 - coordinatae signatae: Punctum/Fines s32; mandata_textus, mandata_coetus_incipere, componens_ponere_transformatio s32; serializatores gemellos attributum_s32/attributum_signatum accipiunt (VI sedes coordinatarum quaeque), v punctorum s32; Eventus.mus i32 manet, semel conversum in destinatio, derivare, manu; custos negativus in invenire SUBLATUS (liberi cum finibus negativis legitimi - umbra probata); pan negativum effusum '-40'. Portae VIII virides, exemplar replay immotum. Culpa plantata: terminus X translationis MUTUS (translationes omnes verticales) - terminus Y rubra ad :148. Custos latinus nihil; formator 0; examen ACCIPE XII plagulis. Worklog (mandatum); consilium II T1 signatum, STATUS.

+++ lib/derivare.c
@@ -57,9 +57,9 @@ derivare (
     }
     alioquin
     {
-        d->tempus_ultimi  = ev->tempus;
-        d->ultimus.x      = ev->datum.mus.x;
-        d->ultimus.y      = ev->datum.mus.y;
-        d->habet_ultimum  = VERUM;
+        d->tempus_ultimi      = ev->tempus;
+                d->ultimus.x  = (s32)ev->datum.mus.x;
+        d->ultimus.y          = (s32)ev->datum.mus.y;
+        d->habet_ultimum      = VERUM;
     }
 }
